requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
supabase==2.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
weasyprint==60.2
plotly==5.17.0
pandas==2.1.3
httpx>=0.24.0,<0.25.0
pydantic[email]==2.5.0
python-decouple==3.8

